feat: Add comment `Fixes #issue` when merging PRs
This adds a comment `Fixes #issue` to the pull request after it's been
merged.

It does this by looking at all the commits and their message and finding
the first mention of `(#nnn)` where `nnn` would be the issue number.

Uses first mention as in a commit message it is expected that only the
first would be relevant to the pull request, but the message might
reference other issues or pull requests via the `#nnn` notation.